[Chore][Core/Dashboard] Merge MetricsHead into APIHead
