+++ python/ray/dashboard/modules/api/metrics/consts.py
@@ -1,0 +1,14 @@
+METRICS_OUTPUT_ROOT_ENV_VAR = "RAY_METRICS_OUTPUT_ROOT"
+DEFAULT_PROMETHEUS_HOST = "http://localhost:9090"
+PROMETHEUS_HOST_ENV_VAR = "RAY_PROMETHEUS_HOST"
+DEFAULT_PROMETHEUS_HEADERS = "{}"
+PROMETHEUS_HEADERS_ENV_VAR = "RAY_PROMETHEUS_HEADERS"
+DEFAULT_PROMETHEUS_NAME = "Prometheus"
+PROMETHEUS_NAME_ENV_VAR = "RAY_PROMETHEUS_NAME"
+PROMETHEUS_HEALTHCHECK_PATH = "-/healthy"
+DEFAULT_GRAFANA_HOST = "http://localhost:3000"
+GRAFANA_HOST_ENV_VAR = "RAY_GRAFANA_HOST"
+GRAFANA_HOST_DISABLED_VALUE = "DISABLED"
+GRAFANA_IFRAME_HOST_ENV_VAR = "RAY_GRAFANA_IFRAME_HOST"
+GRAFANA_DASHBOARD_OUTPUT_DIR_ENV_VAR = "RAY_METRICS_GRAFANA_DASHBOARD_OUTPUT_DIR"
+GRAFANA_HEALTHCHECK_PATH = "api/health"
