+++ python/ray/dashboard/modules/api/metrics/dashboards/data_grafana_dashboard_base.json
