+++ python/ray/dashboard/modules/api/metrics/dashboards/default_grafana_dashboard_base.json
