+++ python/ray/dashboard/modules/api/metrics/templates.py
