+++ python/ray/dashboard/subprocesses/routes.py
@@ -88,6 +88,7 @@ class SubprocessRouteTable(BaseRouteTable):
             # Used in bound_routes().
             parent_side_handler.__route_method__ = method
             parent_side_handler.__route_path__ = path
+            parent_side_handler.__handler__ = handler
 
             cls._routes.route(method, path)(parent_side_handler)
